[pattgen,doc] Pattgen IP Specification
This is the first PR related to getting Pattgen out of Bronze and into master.
Development of the Pattgen IP has been proceeding under the Bronze Branch (all work by @tunghoang290780,
the one-pager author).  This work in bronze includes an initial RTL skeleton, a skeleton of the markdown
documentation (testplan, checklist, etc), and all the hsjon.

From here on out, @tunghoang290780 will be responsible for DV for this IP, and I (@martin-lueker) will be taking on
RTL development.

The first step has been to take all of @tunghoang290780's work and cherry-pick it onto master.
This process has been completed and can be found in the following branch:
https://github.com/martin-lueker/opentitan/tree/pattgen_spec_master

This master-based branch has a mix of RTL, documentation, and some DV related changes.  So from this branch we've
extracted just the documentation (a joint effort by @tunghoang290780 and @martin-lueker) for this initial PR.
At the time of this submission, it is confirmed that these PR contents are enough to render the documentation, using
`util/build_docs.py`

RTL and DV skeletons will follow in a subsequent PR.

Many Thanks to Tung for getting this started, and thanks in advance to our reviewers!
We are looking forward to working with you on this new IP!

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>